fori_loop based on while_loop

+++ test/test_fori_loop_with_while_loop_simple_add_dispatch_in_torch.py
@@ -62,9 +62,9 @@ class WhileLoopTest(unittest.TestCase):
     device = xm.xla_device()
 
     # TODO(@manfei): lower, upper and init_val has to be torch.tensor.
-    init_val = torch.tensor([1], dtype=torch.int32, device=device)
-    lower = torch.tensor([0], dtype=torch.int32, device=device)
-    upper = torch.tensor([10], dtype=torch.int32, device=device)
+    init_val = torch.tensor([1], dtype=torch.float32, device=device)
+    lower = torch.tensor([0], dtype=torch.float32, device=device)
+    upper = torch.tensor([10], dtype=torch.float32, device=device)
 
     def body_fun(init_val):
       return torch.sin(init_val) # init_val)
