+++ torch_xla/experimental/fori_loop.py
@@ -22,7 +22,7 @@ def fori_loop(lower, upper, body_fun, init_val):
   # one_value = torch.tensor([0], dtype=torch.int32, device=device) # torch.ones(1, dtype=torch.int32, device=device)
 
   def cond_fn(lower, upper, init_val):
-    one_value = torch.tensor([0], dtype=torch.int32, device=device) # torch.ones(1, dtype=torch.int32, device=device)
+    one_value = torch.tensor([0], dtype=torch.float32, device=device) # torch.ones(1, dtype=torch.int32, device=device)
     lower = torch.add(lower, one_value)
     return lower[0] >= upper[0]
